extract phone method to a new class
